requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
frozenlist==1.5.0
idna==3.10
joblib==1.4.2
magic-filter==1.0.12
multidict==6.1.0
nltk==3.9.1
propcache==0.2.1
pydantic==2.10.5
pydantic-core==2.27.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
